Move entities models
