+++ tests/bot/models/commands/test_mentions_list.py
@@ -3,8 +3,8 @@ from uuid import UUID, uuid4
 
 import pytest
 
-from botx.bot.models.commands.entities import Mention, MentionList
 from botx.bot.models.commands.enums import MentionTypes
+from botx.models.message.entities import Mention, MentionList
 
 
 @pytest.fixture
